+++ cxparser/ast/program.go
@@ -1,0 +1,8 @@
+package ast
+
+/*
+	Program struct is collection of Statements []Statement.
+*/
+type Program struct {
+	Statements []Statement
+}
