+++ cxparser/prattparser/return.go
@@ -1,0 +1,24 @@
+package parser
+
+import (
+	"github.com/skycoin/cx/cxparser/ast"
+	"github.com/skycoin/cx/cxparser/token"
+)
+
+/*
+	parseReturnStatement returns ast.ReturnStatement.
+*/
+func (p *Parser) parseReturnStatement() *ast.ReturnStatement {
+
+	stmt := &ast.ReturnStatement{Token: p.curToken}
+
+	p.nextToken()
+
+	stmt.ReturnValue = p.parseExpression(LOWEST)
+
+	if p.peekTokenIs(token.SEMICOLON) {
+		p.nextToken()
+	}
+
+	return stmt
+}
